requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0
httpx>=0.23.0
python-dotenv>=0.19.0

# Development dependencies
pytest>=6.2.0
pytest-asyncio>=0.15.0
black>=21.0.0
flake8>=3.8.0